Drop performance profiles; allow maps up to 4096 × 4096
- One area limit (16,777,216 tiles) replaces the light/balanced/high
  profiles; the editor opens 2048 × 2048 unless the browser reports
  less than 8 GB, and warns when a size exceeds 2048 × 2048
- Generate layers row by row instead of per cell (2048²: ~3 s -> 0.3 s,
  4096²: ~1.1 s in Node)
- Stream the TMX as row chunks into Blob batches instead of one string
  (4096² export is ~224 MB)
- Preview cells travel as Uint16Array, halving the transfer
- Presets for 2048 × 2048 and 4096 × 4096

+++ tests/preview.test.mjs
@@ -25,7 +25,7 @@ test('worker transfers previews, recovers from invalid moves and exports the cur
  const call=(kind,args={})=>new Promise((resolve,reject)=>{const id=++serial;worker.once('message',data=>data.error?reject(Error(data.error)):resolve(data.result));worker.postMessage({id,kind,...args});});
  try{
   await call('init',{base});const config=defaults(160,130);
-  const preview=await call('generate',{config});assert.equal(preview.width,160);assert.ok(preview.layers[0].cells instanceof Uint32Array);
+  const preview=await call('generate',{config});assert.equal(preview.width,160);assert.ok(preview.layers[0].cells instanceof Uint16Array);
   const bad=structuredClone(config);bad.Positions.Cave={X:0,Y:0};await assert.rejects(call('generate',{config:bad}));
   config.Positions.Shrine={X:20,Y:7};await call('generate',{config});
   const blob=await call('export',{config});assert.equal(await blob.text(),toTmx(makeMap(base,config)));
